new alarm logic
if alarm is triggered by ALARM CHECK module, the module will update the PICO watchdog with the combined alarm message.  And then the database will wait the alarm to be cleared to have further update.
if there is no alarm in ALARM CHECK module, the DATABASE module will update the current status of database bkg running(runing successfully or not)